Use UTF-8 instead of Latin-1
This decoding happens when a non-unicode string is rendered in
a DTML template.

The decoding happens in 'join_unicode' which takes no arguments. But a
better fix for this would perhaps be to make that function take an encoding
argument (or boolean keyword argument to enable UTF-8).

However, backwards compatibility here seems rather pointless.